+++ examples/server/src/rpc/pub/rabbitmq.ts
@@ -1,14 +1,13 @@
 /* eslint-disable no-console */
 import { createClient } from 'phecda-server/rabbitmq'
 import amqp from 'amqplib'
-import type { TestRpc } from '../test.controller'
-import { TestRpc as Faker } from '../mq'
+import { TestRpc } from '../test.rpc'
 export async function start() {
   const conn = await amqp.connect('amqp://localhost:5672')
 
   const ch = await conn.createChannel()
   const client = await createClient(ch, 'test', {
-    test: Faker as unknown as typeof TestRpc,
+    test: TestRpc,
   })
   const ret = await client.test.run('xx')
   console.log(`return with ${ret}`)