+++ packages/server/register/index.mjs
@@ -1,33 +1,40 @@
-import { register } from "node:module";
-import { MessageChannel } from "node:worker_threads";
-import { log } from "../dist/index.mjs";
-const { port1, port2 } = new MessageChannel();
+import { register } from 'node:module'
+import { MessageChannel } from 'node:worker_threads'
+import { log } from '../dist/index.mjs'
+const { port1, port2 } = new MessageChannel()
 
-register("./loader.mjs", {
+register('./loader.mjs', {
   parentURL: import.meta.url,
   data: { port: port2 },
   transferList: [port2],
-});
+})
 
-port1.on("message", async (data) => {
-  if (!globalThis.__PS_HMR__) return;
+let isRunning = true
 
-  const { type, files } = JSON.parse(data);
-  if (type === "change") {
-    log("reload module...");
+port1.on('message', async (data) => {
+  const { type, files } = JSON.parse(data)
 
-    for (const cb of globalThis.__PS_HMR__) await cb(files);
+  if (!isRunning || !globalThis.__PS_HMR__ || type === 'relaunch')
+    return process.exit(2)// file change -> relaunch
 
-    log("reload done");
+  if (type === 'change') {
+    log('reload module...')
+
+    for (const cb of globalThis.__PS_HMR__) await cb(files)
+
+    log('reload done')
   }
-});
+})
 
 process.on('uncaughtException', (err) => {
-  log('Uncaught Exception:','error')
-  console.error(err);
-});
+  log('Uncaught Exception:', 'error')
+  isRunning = false
+  console.error(err)
+})
 
 process.on('unhandledRejection', (err) => {
-  log('Unhandled Promise Rejection:','error')
-  console.error(err);
-});
+  log('Unhandled Promise Rejection:', 'error')
+  isRunning = false
+
+  console.error(err)
+})