+++ packages/server/register/loader.mjs
@@ -1,206 +1,214 @@
-import { fileURLToPath, pathToFileURL } from "url";
-import { watch, existsSync } from "fs";
-import { writeFile } from "fs/promises";
-import { extname, isAbsolute, relative } from "path";
-import ts from "typescript";
-import { compile, genUnImportRet } from "./utils.mjs";
-import { log, PS_FILE_RE } from "../dist/index.mjs";
+import { fileURLToPath, pathToFileURL } from 'url'
+import { existsSync, watch } from 'fs'
+import { writeFile } from 'fs/promises'
+import { extname, isAbsolute, relative } from 'path'
+import ts from 'typescript'
+import { PS_FILE_RE, log } from '../dist/index.mjs'
+import { compile, genUnImportRet } from './utils.mjs'
 
-let port;
+let port
 
 // this part is important or not?
-const EXTENSIONS = [ts.Extension.Ts, ts.Extension.Tsx, ts.Extension.Mts];
+const EXTENSIONS = [ts.Extension.Ts, ts.Extension.Tsx, ts.Extension.Mts]
 const tsconfig = {
   module: ts.ModuleKind.ESNext,
   moduleResolution: ts.ModuleResolutionKind.NodeNext,
-};
+}
 const moduleResolutionCache = ts.createModuleResolutionCache(
   ts.sys.getCurrentDirectory(),
-  (x) => x,
-  tsconfig
-);
+  x => x,
+  tsconfig,
+)
 const host = {
   fileExists: ts.sys.fileExists,
   readFile: ts.sys.readFile,
-};
+}
 
-let unimportRet;
+let unimportRet
 
-const dtsPath = "ps.d.ts";
+const dtsPath = 'ps.d.ts'
 
 export async function initialize(data) {
-  port = data.port;
+  port = data.port
 
-  if (process.env.PS_NO_DTS) {
-    return;
-  }
+  if (process.env.PS_NO_DTS)
+    return
 
-  unimportRet = await genUnImportRet();
+  unimportRet = await genUnImportRet()
 
   if (unimportRet) {
-    log("auto import...");
+    log('auto import...')
     if (!existsSync(dtsPath))
-      writeFile(dtsPath, await unimportRet.generateTypeDeclarations());
+      writeFile(dtsPath, await unimportRet.generateTypeDeclarations())
   }
 }
 
-const watchFiles = new Set();
-const filesRecord = new Map();
-const moduleGraph = {};
+const watchFiles = new Set()
+const filesRecord = new Map()
+const moduleGraph = {}
 
-let entryUrl;
+let entryUrl
 
 function addUrlToGraph(url, parent) {
-  if (!(url in moduleGraph)) moduleGraph[url] = new Set();
+  if (!(url in moduleGraph))
+    moduleGraph[url] = new Set()
 
-  moduleGraph[url].add(parent);
-  return url + (filesRecord.has(url) ? `?t=${filesRecord.get(url)}` : "");
+  moduleGraph[url].add(parent)
+  return url + (filesRecord.has(url) ? `?t=${filesRecord.get(url)}` : '')
 }
 
 export const resolve = async (specifier, context, nextResolve) => {
   // entrypoint
   if (!context.parentURL) {
-    entryUrl = specifier;
+    entryUrl = specifier
     return {
-      format: EXTENSIONS.some((ext) => specifier.endsWith(ext))
-        ? "ts"
+      format: EXTENSIONS.some(ext => specifier.endsWith(ext))
+        ? 'ts'
         : undefined,
       url: specifier,
       shortCircuit: true,
-    };
+    }
   }
-  if (/^file:\/\/\//.test(specifier) && extname(specifier) === ".ts") {
-    const url = addUrlToGraph(specifier, context.parentURL.split("?")[0]);
+  if (/^file:\/\/\//.test(specifier) && extname(specifier) === '.ts') {
+    const url = addUrlToGraph(specifier, context.parentURL.split('?')[0])
 
     return {
-      format: "ts",
+      format: 'ts',
       url,
       shortCircuit: true,
-    };
+    }
   }
 
   if (
-    context.parentURL.includes("/node_modules/phecda-server") &&
-    isAbsolute(specifier)
+    context.parentURL.includes('/node_modules/phecda-server')
+    && isAbsolute(specifier)
   ) {
     specifier = relative(fileURLToPath(entryUrl), specifier)
-      .replace(/\.ts$/, "")
-      .slice(1);
-    context.parentURL = entryUrl;
+      .replace(/\.ts$/, '')
+      .slice(1)
+    context.parentURL = entryUrl
   }
   // import/require from external library
-  if (context.parentURL.includes("/node_modules/"))
-    return nextResolve(specifier);
+  if (context.parentURL.includes('/node_modules/'))
+    return nextResolve(specifier)
 
   const { resolvedModule } = ts.resolveModuleName(
     specifier,
     fileURLToPath(context.parentURL),
     tsconfig,
     host,
-    moduleResolutionCache
-  );
+    moduleResolutionCache,
+  )
   // import from local project to local project TS file
   if (
-    resolvedModule &&
-    !resolvedModule.resolvedFileName.includes("/node_modules/") &&
-    EXTENSIONS.includes(resolvedModule.extension)
+    resolvedModule
+    && !resolvedModule.resolvedFileName.includes('/node_modules/')
+    && EXTENSIONS.includes(resolvedModule.extension)
   ) {
     const url = addUrlToGraph(
       pathToFileURL(resolvedModule.resolvedFileName).href,
-      context.parentURL.split("?")[0]
-    );
+      context.parentURL.split('?')[0],
+    )
 
     return {
-      format: "ts",
+      format: 'ts',
       url,
       shortCircuit: true,
-    };
+    }
   }
-  return nextResolve(specifier);
-};
+  return nextResolve(specifier)
+}
 
 export const load = async (url, context, nextLoad) => {
-  url = url.split("?")[0];
+  url = url.split('?')[0]
   if (
-    !url.includes("/node_modules/") &&
-    url.startsWith("file://") &&
-    !watchFiles.has(url)
+    !url.includes('/node_modules/')
+    && url.startsWith('file://')
+    && !watchFiles.has(url)
   ) {
-    watchFiles.add(url);
-
+    watchFiles.add(url)
     watch(
       fileURLToPath(url),
       debounce((type) => {
-        if (type === "change") {
+        if (type === 'change') {
           try {
-            const files = [...findTopScope(url, Date.now())].reverse();
+            const files = [...findTopScope(url, Date.now())].reverse()
+
             port.postMessage(
               JSON.stringify({
-                type: "change",
+                type: 'change',
                 files,
-              })
-            );
-          } catch (e) {
-            process.exit(3);
+              }),
+            )
+          }
+          catch (e) {
+            port.postMessage(
+              JSON.stringify({
+                type: 'relaunch',
+              }),
+            )
           }
         }
-      })
-    );
+      }),
+    )
   }
-  //resolveModuleName failed
+  // resolveModuleName failed
   // I don't know why it failed
-  if (!context.format && url.endsWith(".ts")) {
-    context.format = "ts";
-  }
+  if (!context.format && url.endsWith('.ts'))
+    context.format = 'ts'
 
-  if (context.format === "ts") {
-    const { source } = await nextLoad(url, context);
+  if (context.format === 'ts') {
+    const { source } = await nextLoad(url, context)
 
-    const code =
-      typeof source === "string" ? source : Buffer.from(source).toString();
-    const compiled = await compile(code, url);
-    if (unimportRet && isModuleFileUrl(url)) {
-      const { injectImports } = unimportRet;
+    const code
+      = typeof source === 'string' ? source : Buffer.from(source).toString()
+    const compiled = await compile(code, url)
+    if (unimportRet) {
+      const { injectImports } = unimportRet
       return {
-        format: "module",
+        format: 'module',
         source: (await injectImports(compiled)).code,
         shortCircuit: true,
-      };
+      }
     }
     return {
-      format: "module",
+      format: 'module',
       source: compiled,
       shortCircuit: true,
-    };
-  } else {
-    return nextLoad(url, context);
+    }
   }
-};
+  else {
+    return nextLoad(url, context)
+  }
+}
 
 function findTopScope(url, time, modules = new Set()) {
-  filesRecord.set(url, time);
+  filesRecord.set(url, time)
   if (isModuleFileUrl(url)) {
-    modules.add(fileURLToPath(url));
-  } else {
-    if (!moduleGraph[url]) throw new Error("root file update");
-    for (const i of [...moduleGraph[url]]) findTopScope(i, time, modules);
+    modules.add(fileURLToPath(url))
+  }
+  else {
+    if (!moduleGraph[url])
+      throw new Error('root file update')
+    for (const i of [...moduleGraph[url]]) findTopScope(i, time, modules)
   }
 
-  return modules;
+  return modules
 }
 
 function debounce(cb, timeout = 500) {
-  let timer;
+  let timer
   return (...args) => {
-    if (timer) return;
+    if (timer)
+      return
 
     timer = setTimeout(() => {
-      cb(...args);
-      timer = undefined;
-    }, timeout);
-  };
+      cb(...args)
+      timer = undefined
+    }, timeout)
+  }
 }
 
 export function isModuleFileUrl(url) {
-  return PS_FILE_RE.test(url);
+  return PS_FILE_RE.test(url)
 }